Proposal for how to do TypedNodes and correctly subclass Node
