+++ cognite/client/data_classes/data_modeling/__init__.py
@@ -76,15 +76,13 @@ from cognite.client.data_classes.data_modeling.instances import (
     NodeList,
     NodeListWithCursor,
     NodeOrEdgeData,
-)
-from cognite.client.data_classes.data_modeling.spaces import Space, SpaceApply, SpaceApplyList, SpaceList
-from cognite.client.data_classes.data_modeling.typed_instances import (
     PropertyOptions,
     TypedEdge,
     TypedEdgeApply,
     TypedNode,
     TypedNodeApply,
 )
+from cognite.client.data_classes.data_modeling.spaces import Space, SpaceApply, SpaceApplyList, SpaceList
 from cognite.client.data_classes.data_modeling.views import (
     ConnectionDefinition,
     EdgeConnection,
