+++ tests/tests_unit/test_data_classes/test_data_models/test_typed_instances.py
@@ -5,13 +5,19 @@ from typing import Any
 
 import pytest
 
-from cognite.client.data_classes.data_modeling import DirectRelationReference, NodeList, ViewId
+from cognite.client.data_classes.data_modeling import (
+    DirectRelationReference,
+    NodeList,
+    NodeOrEdgeData,
+    ViewId,
+)
 from cognite.client.data_classes.data_modeling.cdm.v1 import (
     CogniteAsset,
     CogniteAssetApply,
     CogniteDescribableEdgeApply,
 )
-from cognite.client.data_classes.data_modeling.typed_instances import (
+from cognite.client.data_classes.data_modeling.instances import (
+    Properties,
     PropertyOptions,
     TypedEdge,
     TypedEdgeApply,
@@ -27,13 +33,24 @@ class PersonProperties:
 class Person(TypedNodeApply, PersonProperties):
     def __init__(
         self,
+        space: str,
         external_id: str,
         name: str,
         birth_date: date,
         email: str,
         siblings: list[DirectRelationReference] | None = None,
     ) -> None:
-        super().__init__("sp_my_fixed_space", external_id, type=DirectRelationReference("sp_model_space", "person"))
+        super().__init__(
+            space,
+            external_id,
+            type=DirectRelationReference("sp_model_space", "person"),
+            sources=[
+                NodeOrEdgeData(
+                    source=self.get_source(),
+                    properties={"birthDate": birth_date, "email": email, "name": name, "siblings": siblings or []},
+                )
+            ],
+        )
         self.name = name
         self.birth_date = birth_date
         self.email = email
@@ -57,7 +74,12 @@ class FlowWrite(TypedEdgeApply, FlowProperties):
         flow_rate: float,
     ) -> None:
         super().__init__(
-            "sp_my_fixed_space", external_id, DirectRelationReference("sp_model_space", "Flow"), start_node, end_node
+            "sp_my_fixed_space",
+            external_id,
+            DirectRelationReference("sp_model_space", "Flow"),
+            start_node,
+            end_node,
+            sources=[NodeOrEdgeData(source=self.get_source(), properties={"flowRate": flow_rate})],
         )
         self.flow_rate = flow_rate
 
@@ -81,7 +103,16 @@ class Flow(TypedEdge, FlowProperties):
         deleted_time: int | None,
     ) -> None:
         super().__init__(
-            space, external_id, version, type, last_updated_time, created_time, start_node, end_node, deleted_time, None
+            space,
+            external_id,
+            version,
+            type,
+            last_updated_time,
+            created_time,
+            start_node,
+            end_node,
+            deleted_time,
+            properties=Properties({self.get_source(): {"flowRate": flow_rate}}),
         )
         self.flow_rate = flow_rate
 
@@ -105,14 +136,31 @@ class PersonRead(TypedNode, PersonProperties):
         type: DirectRelationReference | tuple[str, str] | None = None,
         deleted_time: int | None = None,
     ) -> None:
-        super().__init__(space, external_id, version, last_updated_time, created_time, deleted_time, None, type)
+        super().__init__(
+            space,
+            external_id,
+            version,
+            last_updated_time,
+            created_time,
+            deleted_time,
+            properties=Properties.from_typed_properties(
+                self.get_source(),
+                {
+                    "name": name,
+                    "birthDate": birth_date,
+                    "email": email,
+                    "siblings": siblings,
+                },
+            ),
+            type=type,
+        )
         self.name = name
         self.birth_date = birth_date
         self.email = email
         self.siblings = siblings
 
     def as_write(self) -> Person:
-        return Person(self.external_id, self.name, self.birth_date, self.email, self.siblings)
+        return Person(self.space, self.external_id, self.name, self.birth_date, self.email, self.siblings)
 
     @classmethod
     def get_source(cls) -> ViewId:
@@ -123,7 +171,12 @@ class Asset(TypedNodeApply):
     type_ = PropertyOptions(identifier="type")
 
     def __init__(self, external_id: str, name: str, type_: str) -> None:
-        super().__init__("sp_my_fixed_space", external_id, type=DirectRelationReference("sp_model_space", "asset"))
+        super().__init__(
+            "sp_my_fixed_space",
+            external_id,
+            type=DirectRelationReference("sp_model_space", "asset"),
+            sources=[NodeOrEdgeData(source=self.get_source(), properties={"name": name, "type": type_})],
+        )
         self.name = name
         self.type_ = type_
 
@@ -134,7 +187,7 @@ class Asset(TypedNodeApply):
 
 class TestTypedNodeApply:
     def test_dump_person(self) -> None:
-        person = Person("my_external_id", "John Doe", date(1990, 1, 1), "example@cognite.com")
+        person = Person("sp_my_fixed_space", "my_external_id", "John Doe", date(1990, 1, 1), "example@cognite.com")
         expected = {
             "space": "sp_my_fixed_space",
             "externalId": "my_external_id",
@@ -143,7 +196,12 @@ class TestTypedNodeApply:
             "sources": [
                 {
                     "source": {"space": "sp_model_space", "externalId": "view_id", "version": "1", "type": "view"},
-                    "properties": {"name": "John Doe", "birthDate": "1990-01-01", "email": "example@cognite.com"},
+                    "properties": {
+                        "name": "John Doe",
+                        "birthDate": "1990-01-01",
+                        "email": "example@cognite.com",
+                        "siblings": [],
+                    },
                 }
             ],
         }
@@ -233,7 +291,8 @@ class TestTypedNode:
             },
         }
 
-        assert person.dump() == expected
+        actual = person.dump()
+        assert actual == expected
         loaded = PersonRead.load(expected)
         assert person == loaded
         assert isinstance(loaded.birth_date, date)
@@ -256,8 +315,9 @@ class TestTypedNode:
                 "created_time": pd.Timestamp("1970-01-01"),
                 "instance_type": "node",
                 "name": "John Doe",
-                "birth_date": "1990-01-01",
+                "birthDate": "1990-01-01",
                 "email": "john@doe.com",
+                "siblings": None,
             },
         ).to_frame(name="value")
         pd.testing.assert_frame_equal(df, expected_df)
@@ -285,6 +345,7 @@ class TestTypedNode:
                     "name": ["John Doe"],
                     "birthDate": ["1990-01-01"],
                     "email": ["john@doe.com"],
+                    "siblings": None,
                 }
             ),
         )
